# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_interfaces_http_handlers.CreateMoodRequest:
    properties:
      date:
        type: string
      intensity:
        maximum: 10
        minimum: 1
        type: integer
      is_shared:
        type: boolean
      mood_type:
        allOf:
        - $ref: '#/definitions/loviary_app_backend_internal_domain_moods.MoodType'
        enum:
        - happy
        - sad
        - angry
        - anxious
        - calm
        - excited
        - tired
        - love
        - stressed
        - grateful
      note:
        type: string
    required:
    - date
    - intensity
    - mood_type
    type: object
...
  internal_interfaces_http_handlers.ErrorResponse:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: validation failed
        type: string
    type: object
...
  internal_interfaces_http_handlers.MoodListResponse:
    properties:
      count:
        example: 5
        type: integer
      moods:
        items:
          $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.MoodResponse'
        type: array
    type: object
...
  internal_interfaces_http_handlers.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  internal_interfaces_http_handlers.UpdateMoodRequest:
    properties:
      intensity:
        maximum: 10
        minimum: 1
        type: integer
      is_shared:
        type: boolean
      mood_type:
        allOf:
        - $ref: '#/definitions/loviary_app_backend_internal_domain_moods.MoodType'
        enum:
        - happy
        - sad
        - angry
        - anxious
        - calm
        - excited
        - tired
        - love
        - stressed
        - grateful
      note:
        type: string
    type: object
...
  loviary_app_backend_internal_domain_moods.MoodType:
    enum:
    - happy
    - sad
    - angry
    - anxious
    - calm
    - excited
    - tired
    - love
    - stressed
    - grateful
    type: string
    x-enum-varnames:
    - MoodTypeHappy
    - MoodTypeSad
    - MoodTypeAngry
    - MoodTypeAnxious
    - MoodTypeCalm
    - MoodTypeExcited
    - MoodTypeTired
    - MoodTypeLove
    - MoodTypeStressed
    - MoodTypeGrateful
...
  loviary_app_backend_internal_interfaces_http_dto.MoodResponse:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      intensity:
        type: integer
      is_shared:
        type: boolean
      mood_type:
        $ref: '#/definitions/loviary_app_backend_internal_domain_moods.MoodType'
      note:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@loviary.app
    name: API Support
  description: Loviary - Couple Relationship Tracking API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Loviary API
  version: "1.0"
paths:
...
  /moods:
    post:
      consumes:
      - application/json
      description: Log a new mood entry with type and intensity
      parameters:
      - description: Mood data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.CreateMoodRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Mood created
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.MoodResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create mood
      tags:
      - moods
  /moods/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a mood entry by ID
      parameters:
      - description: Mood ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mood deleted
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.SuccessResponse'
        "400":
          description: Invalid mood ID
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Mood not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete mood
      tags:
      - moods
    get:
      consumes:
      - application/json
      description: Get mood entry by ID
      parameters:
      - description: Mood ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mood data
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.MoodResponse'
        "400":
          description: Invalid mood ID
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Mood not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get mood
      tags:
      - moods
    patch:
      consumes:
      - application/json
      description: Update an existing mood entry
      parameters:
      - description: Mood ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated mood data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.UpdateMoodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated mood
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.MoodResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Mood not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update mood
      tags:
      - moods
  /moods/history:
    get:
      consumes:
      - application/json
      description: Get mood history with optional date range filtering
      parameters:
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mood history with count
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.MoodListResponse'
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get mood history
      tags:
      - moods
  /moods/today:
    get:
      consumes:
      - application/json
      description: Get mood logged for today
      produces:
      - application/json
      responses:
        "200":
          description: Today's mood
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.MoodResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: No mood logged today
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get today's mood
      tags:
      - moods
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"